Update RegistrationTests.py
- Defining tests for RegistrationTests from the RegistrationTests.feature file

At this point, I think most of the framework for testing this application is complete. I think from this point onward, it really will be repeating actions: define tests in Gherkin language and write them in feature file, create test file, and implement tests while filling out more of the pages as needed. Because it is repeatable, I might only do some as more of an example and purposely leave the rest. This is to show what I know in regards to web browser/GUI automation. There are basically an infinite number of tests I could write for this tiny web app. No need to do it all, just need to show I can.