Implemented NewUserRegisters Test
- filled out test_NewUserRegisters test method following feature file writeup
- added more wait methods to Element.py. I will cleanup/consolidate these next commit
-created User.py object, which represents a user and associated information. This will allow to take hardcoded username/passwords out of Login tests. I'll use the BookStoreApi to create new users and delete new users for each test.
- filled out RegisterPage.py with elements and methods.

The RegisterPage has a captcha element, which makes it hard for automation tests to run through the registration process through the UI. There are a few solutions to this: 1. put a sleep call in the test to have the tester solve the captcha manually (implemented here) 2. to use cookies from 1 solved test to run other tests with until the cookie expires or 3. if I had access to the testing environment of the web app, to use the specific key to allow for phony captchas to be put into place. I ended up implementing the first method because it was quickest and easiest for the demo after attempting the second method real quick. In a testing environment, option 3 would apply.

Next commit is a code cleanup and then I will add the rest of the tests for this demo.